refactor: ReservationTimeAvailability를 repository의 dto로 이동
