+++ src/main/java/roomescape/repository/ReservationTimeRepository.java
@@ -1,7 +1,7 @@
 package roomescape.repository;
 
 import roomescape.domain.ReservationTime;
-import roomescape.service.dto.ReservationTimeAvailability;
+import roomescape.repository.dto.ReservationTimeAvailability;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
