+++ src/main/java/roomescape/repository/dto/ReservationTimeAvailability.java
@@ -1,4 +1,4 @@
-package roomescape.service.dto;
+package roomescape.repository.dto;
 
 
 import lombok.Getter;
