+++ src/test/java/roomescape/repository/JdbcReservationTimeRepositoryTest.java
@@ -8,7 +8,7 @@ import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.context.annotation.Import;
 import roomescape.domain.Reservation;
 import roomescape.domain.ReservationTime;
-import roomescape.service.dto.ReservationTimeAvailability;
+import roomescape.repository.dto.ReservationTimeAvailability;
 import roomescape.domain.Theme;
 
 import java.time.LocalDate;
